refactoring part 3
