+++ Sources/RSocketCore/Extensions/Coder/DecoderBuilder.swift
@@ -1,0 +1,44 @@
+/*
+ * Copyright 2015-present the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+@resultBuilder
+public enum DecoderBuilder: DecoderBuilderProtocol {
+    public static func buildBlock<Decoder>(
+        _ decoder: Decoder
+    ) -> Decoder where Decoder: DecoderProtocol {
+        decoder
+    }
+}
+
+public protocol DecoderBuilderProtocol {}
+
+public extension DecoderBuilderProtocol {
+    static func buildExpression<Decoder>(
+        _ decoder: Decoder
+    ) -> Decoders.PreserveMetadata<Decoder> {
+        decoder.preserveMetadata()
+    }
+    static func buildExpression<Decoder>(
+        _ decoder: Decoder
+    ) -> Decoder where Decoder: DecoderProtocol, Decoder.Metadata == Void {
+        decoder
+    }
+    static func buildExpression<Decoder>(
+        _ decoder: Decoder
+    ) -> Decoders.EraseMetadata<Decoder> where Decoder.Metadata: ErasableMetadata {
+        decoder.eraseMetadata()
+    }
+}
